Update Eustathios_Spider_V2.5.f3d
I added Ridged Bodies, Grounding of the frame, and Joints for motion study to the Fusion360 Model. This took me longer than expected because this is an imported model and I had to combine, create components, and rearrange the model structure from Solidworks to how it would work best in Fusion360.