Added clean 1:1 DXF prints for heated bed and acrylic panels
The DXF detailed prints with title blocks are kind of messy for importing to some cutting programs. So I cleaned them up and scaled them 1:1 in the DXF so they are easy to send to a laser cutter (Acrylic panels), or to program a simple cut on a mill to cut/drill the Cast Aluminum Heated Bed. I used layers so that each feature type is specified and are easy to turn on and off.